split helper func into own file
